services/vectorization/requirements.txt
# Vectorization Service Dependencies
fastapi
uvicorn[standard]
asyncpg
openai
tiktoken
pgvector
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
numpy
aiohttp
requests